bot/pyproject.toml
[tool.poetry]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
gigachain-community = "^0.2.16.post3"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
